+++ src/main/java/cart/controller/ProductResourceController.java
@@ -1,50 +1,29 @@
 package cart.controller;
 
-import cart.domain.Product;
 import cart.dto.ProductDto;
 import cart.request.ProductRequest;
 import cart.service.ProductService;
-import java.util.List;
 import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
-public class ProductController {
+@RestController
+public class ProductResourceController {
 
     private final ProductService productService;
 
-    public ProductController(final ProductService productService) {
+    public ProductResourceController(final ProductService productService) {
         this.productService = productService;
     }
 
-    @GetMapping("/")
-    public String indexPage(final Model model) {
-        final List<Product> products = productService.findAll();
-
-        model.addAttribute("products", products);
-
-        return "index";
-    }
-
-    @GetMapping("/admin")
-    public String adminPage(final Model model) {
-        final List<Product> products = productService.findAll();
-
-        model.addAttribute("products", products);
-
-        return "admin";
-    }
-
     @PostMapping("/admin")
+    @ResponseStatus(value = HttpStatus.CREATED)
     public void createProduct(@RequestBody @Valid final ProductRequest productRequest) {
         final ProductDto productDto = new ProductDto(
                 productRequest.getName(),
